added a services folder. refactored router files to make it independent from db querying. Still need to add a basic login auth to test with frontend.
