+++ app-backend/api/models/Devices.py
@@ -5,7 +5,7 @@ class Devices(SQLModel, table=True):
 
     deviceId: int = Field(default=None, primary_key=True)
     buildingId: int = Field(index=True, foreign_key="Buildings.buildingId")
-    type: str 
+    deviceTypeId: str 
     status: str = Field(index=True)
     positionX: int 
     positionY: int
