+++ app-backend/api/routes/devices.py
@@ -1,10 +1,8 @@
-from fastapi import APIRouter, HTTPException, Depends
-from sqlmodel import Session, select
-from typing import Annotated
+from fastapi import APIRouter, HTTPException
 
 from ..models.response_models.DevicePosition import DevicePosition
-from ..models.Devices import Devices
-from ..db.engine import get_engine
+from ..db.session import SessionDep
+from ..services.devices import *
 
 router = APIRouter(
     prefix="/devices",
@@ -12,19 +10,11 @@ router = APIRouter(
     responses={404: {"description": "Not found"}},
 )
 
-engine = get_engine()
-
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-SessionDep = Annotated[Session, Depends(get_session)]
 
 @router.get("/")
 def get_devices(session: SessionDep):
     try:
-        statement = select(Devices)
-        devices = session.exec(statement).all()
+        devices = fetch_devices(session)
         return devices
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error retrieving devices: {e}")
@@ -33,8 +23,7 @@ def get_devices(session: SessionDep):
 @router.get("/device-positions",  response_model=list[DevicePosition])
 def get_device_positions(session: SessionDep, limit: int = 100):
     try:
-        statement = select(Devices.deviceId, Devices.positionX, Devices.positionY).limit(limit)
-        positions = session.exec(statement).all()
+        positions = fetch_device_positions(session, limit)
         return [DevicePosition(deviceId=dId, positionX=x, positionY=y) for dId, x, y in positions]
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error retrieving device positions: {e}")
